Cleaned up code, abstracted unneeded shit away from Font class
